refactor(fundamentalanalysisresult.cs): improve code health with declarative rules
FundamentalAnalysisResult captures various fundamentals. It did do this using a series of
if-conditions. This imperative programming technique mixes what (the rules) from how (control flow).
This adds complexity. One easy way to detect this is with a code complexity measure such as
cyclomatic complexity. Using if-conditions increases this complexity. The code is now using a
dictionary to contain the rules which are processed more functionally.

closes #6